+++ lib/mongoid/relations/embedded/many.rb
@@ -133,8 +133,12 @@ module Mongoid
           doc = target.delete_one(document)
           _unscoped.delete_one(doc)
           if doc && !_binding?
-            if _assigning? && !doc.paranoid?
-              base.add_atomic_pull(doc)
+            if _assigning?
+              if doc.paranoid?
+                doc.destroy(suppress: true)
+              else
+                base.add_atomic_pull(doc)
+              end
             else
               doc.delete(suppress: true)
             end