+++ CHANGELOG.md
@@ -789,6 +789,9 @@ For instructions on upgrading to newer versions, visit
 * \#2003 Don't fail on document generation when an embedded document was
   stored as nil in the database.
 
+* \#1997 Don't delete paranoid embedded docs via nested attributes when
+  a before_destroy callback returns false.
+
 * \#1994 `dependent: :delete` only hits the database once now for one to
   many and many to many relations instead of once for each document.